+++ examples/apps/todomvc/src/App.jsx
@@ -1,20 +1,13 @@
 import "./style.css"
 
-import { Provider } from "react-redux"
-
 import Footer from "./Footer"
 import Form from "./Form"
 import List from "./List"
-import { store } from "./store"
-
-const ONE_SECOND = 1000
-const FPS = 30
-
-setInterval(() => store.update(), ONE_SECOND / FPS)
+import { Provider } from "./store"
 
 export default function App() {
   return (
-    <Provider store={store}>
+    <Provider>
       <Form />
       <List />
       <Footer />