+++ examples/apps/todomvc/src/Form.jsx
@@ -1,17 +1,15 @@
-import { useDispatch, useSelector } from "react-redux"
-
+import { useNotify, useSelector } from "./store"
 import { selectValue } from "./store/selectors"
 
 export default function Form() {
-  const dispatch = useDispatch()
+  const notify = useNotify()
 
   const value = useSelector(selectValue)
-  const handleChange = (event) =>
-    dispatch({ type: "inputChange", payload: event.target.value })
+  const handleChange = (event) => notify("inputChange", event.target.value)
 
   const handleSubmit = (event) => {
     event.preventDefault()
-    dispatch({ type: "formSubmit", payload: value })
+    notify("formSubmit", value)
   }
 
   return (