+++ examples/apps/todomvc/src/store/selectors.js
@@ -1,12 +1,12 @@
 import { createSelector } from "@inglorious/store/select"
 
-export const selectValue = ({ entities }) => entities.form.value
-export const selectTasks = ({ entities }) => entities.list.tasks
+export const selectValue = (entities) => entities.form.value
+export const selectTasks = (entities) => entities.list.tasks
 
 export const selectTasksCount = (filter) =>
   createSelector([selectTasks], (tasks) => getTasks(tasks, filter).length)
 
-export const selectActiveFilter = ({ entities }) => entities.footer.activeFilter
+export const selectActiveFilter = (entities) => entities.footer.activeFilter
 
 export const selectFilteredTasks = createSelector(
   [selectTasks, selectActiveFilter],