+++ packages/renderer-react-dom/src/game/game.jsx
@@ -25,8 +25,8 @@ const Z = 2
 
 export default function Game({ engine }) {
   // NOTE: don't use simply engine.entities here: need to subscribe to animate scene!
-  const entities = useSelector((state) => state.entities)
-  // const entities = engine._store.getState().entities
+  const entities = useSelector((state) => state)
+  // const entities = engine._store.getState()
 
   const types = engine._store.getTypes()
   const { mouse, ...rest } = entities